chore(deps): upgrade all frontend and backend dependencies to latest versions
Upgrade the full dependency tree for both the Tauri frontend (npm) and
Rust backend (Cargo) to their latest major/minor versions, and fix all
resulting breaking changes so the project compiles and builds cleanly.

Frontend (npm) major upgrades:
- react/react-dom 18 -> 19
- antd 5 -> 6 (Steps.Step removed, migrated to items prop)
- @ant-design/icons 5 -> 6
- react-router-dom 6 -> 7
- echarts 5 -> 6
- zustand 4 -> 5
- vite 5 -> 8 (now requires esbuild as separate dev dependency)
- tailwindcss 3 -> 4 (PostCSS plugin moved to @tailwindcss/postcss;
  CSS-based config replaces tailwind.config.js)
- typescript 5 -> 7
- @vitejs/plugin-react 4 -> 6
- Vite build target updated from safari13 to safari16 to support
  destructuring syntax used by modern libraries

Backend (Cargo) major upgrades:
- rusqlite 0.32 -> 0.40 (no longer implements FromSql for u64)
- thiserror 1 -> 2
- reqwest 0.12 -> 0.13
- regex 1.12.3 -> 1.13.1

Type and safety fixes:
- Change StockQuote.volume from u64 to i64 to match SQLite's native
  INTEGER type, eliminating lossy as-i64 / as-u64 casts on both the
  insert and select paths. All volume-related conversions across
  Yahoo, Xueqiu, and database layers updated accordingly.

Breaking changes handled:
- antd v6: Steps.Step -> Steps items prop (Import page)
- Tailwind v4: postcss.config.js updated to use @tailwindcss/postcss;
  global.css migrated from @tailwind directives to @import "tailwindcss";
  tailwind.config.js removed (CSS-based config in Tailwind v4)
- Vite 8: esbuild added as devDependency for legacy transform support
- rusqlite 0.40: u64 FromSql removed, volume read as i64 natively

+++ src-tauri/src/services/quote_service.rs
@@ -198,7 +198,7 @@ pub fn save_quotes_to_db(db: &Database, quotes: &[StockQuote]) -> Result<(), Str
             q.change_percent,
             q.high,
             q.low,
-            q.volume as i64,
+            q.volume,
             q.updated_at,
         ])
         .map_err(|e| e.to_string())?;
@@ -376,7 +376,7 @@ pub async fn fetch_yahoo_quote(symbol: &str, market: &str) -> Result<StockQuote,
         .and_then(|q| q.volume.as_ref())
         .and_then(|v| v.last())
         .and_then(|v| *v)
-        .unwrap_or(0);
+        .unwrap_or(0) as i64;
 
     let name = meta
         .short_name
@@ -686,7 +686,7 @@ fn parse_eastmoney_quote(symbol: &str, market: &str, resp: EastMoneyResponse) ->
 
     let high = data.f44.unwrap_or(0.0);
     let low = data.f45.unwrap_or(0.0);
-    let volume = data.f47.unwrap_or(0.0) as u64;
+    let volume = data.f47.unwrap_or(0.0) as i64;
 
     Ok(StockQuote {
         symbol: symbol.to_string(),
@@ -1071,7 +1071,7 @@ fn parse_xueqiu_quote(symbol: &str, market: &str, resp: XueqiuResponse) -> Resul
 
     let high = quote.high.unwrap_or(0.0);
     let low = quote.low.unwrap_or(0.0);
-    let volume = quote.volume.unwrap_or(0.0) as u64;
+    let volume = quote.volume.unwrap_or(0.0) as i64;
 
     Ok(StockQuote {
         symbol: symbol.to_string(),
